pyproject.toml
[project]
name = "retail-media-creative-builder"
version = "2.0.0"
description = "AI-Powered Advertising Creative Generation Platform"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "Flask>=3.0.0",
    "Flask-CORS>=4.0.0",
    "Werkzeug>=3.0.0",
    "google-genai",
    "rembg>=2.0.0",
    "onnxruntime>=1.16.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0,<2.0.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "tqdm>=4.66.0",
    "colorama>=0.4.6",
    "PyYAML>=6.0.0",
    "piexif>=1.1.3",
    "filetype>=1.2.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["backend"]
